+++ DependencyInjection/c975LEmailExtension.php
@@ -17,9 +17,6 @@ use Symfony\Component\DependencyInjection\Loader\YamlFileLoader;
 
 class c975LEmailExtension extends Extension
 {
-    /**
-     * {@inheritdoc}
-     */
     public function load(array $configs, ContainerBuilder $container)
     {
         $loader = new YamlFileLoader(